Fix issue when parsing %-based bg position
As seen in #89, in some browser an issue happens when plugin first fetches bg position. If browser returns '50%', parseInt interprets it as 50 and script later thinks it's 50px.

In that case, it's best to fall back to 0, which this commit does by checking for existence of '%' in the returned bg positions.